Simplyfy Cloak and Manifest Effects (#4652)
