+++ forge-ai/src/main/java/forge/ai/ability/CloakAi.java
@@ -1,58 +1,15 @@
 package forge.ai.ability;
 
-import java.util.Map;
-
-import com.google.common.base.Predicate;
-import com.google.common.collect.Iterables;
-
 import forge.ai.ComputerUtil;
-import forge.ai.ComputerUtilCard;
-import forge.ai.SpellAbilityAi;
-import forge.game.Game;
 import forge.game.card.Card;
-import forge.game.card.CardCollection;
-import forge.game.card.CardLists;
-import forge.game.card.CardPredicates;
 import forge.game.card.CardUtil;
 import forge.game.player.Player;
-import forge.game.player.PlayerActionConfirmMode;
 import forge.game.spellability.SpellAbility;
-import forge.game.zone.ZoneType;
 
-public class CloakAi extends SpellAbilityAi {
-
-    @Override
-    public boolean confirmAction(Player player, SpellAbility sa, PlayerActionConfirmMode mode, String message, Map<String, Object> params) {
-        return true;
-    }
+public class CloakAi extends ManifestBaseAi {
 
     @Override
-    protected boolean checkApiLogic(final Player ai, final SpellAbility sa) {
-        final Game game = ai.getGame();
-        final Card host = sa.getHostCard();
-        if (ComputerUtil.preventRunAwayActivations(sa)) {
-            return false;
-        }
-
-        if (sa.hasParam("Choices") || sa.hasParam("ChoiceZone")) {
-            ZoneType choiceZone = ZoneType.Hand;
-            if (sa.hasParam("ChoiceZone")) {
-                choiceZone = ZoneType.smartValueOf(sa.getParam("ChoiceZone"));
-            }
-            CardCollection choices = new CardCollection(game.getCardsIn(choiceZone));
-            if (sa.hasParam("Choices")) {
-                choices = CardLists.getValidCards(choices, sa.getParam("Choices"), ai, host, sa);
-            }
-            // cloak up only for creatures?
-            choices = CardLists.filter(choices, CardPredicates.isType("Creature"));
-            if (choices.isEmpty()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    static boolean shouldCloak(final Card card, final Player ai, final SpellAbility sa) {
+    protected boolean shouldApply(final Card card, final Player ai, final SpellAbility sa) {
         // check to ensure that there are no replacement effects that prevent creatures ETBing from library
         // (e.g. Grafdigger's Cage)
         Card topCopy = CardUtil.getLKICopy(card);
@@ -82,24 +39,5 @@ public class CloakAi extends SpellAbilityAi {
             }
         }
         return true;
-    }
-
-    @Override
-    protected Card chooseSingleCard(final Player ai, final SpellAbility sa, Iterable<Card> options, boolean isOptional, Player targetedPlayer, Map<String, Object> params) {
-        if (Iterables.size(options) > 1 || isOptional) {
-            CardCollection filtered = CardLists.filter(options, new Predicate<Card>() {
-                @Override
-                public boolean apply(Card input) {
-                    return shouldCloak(input, ai, sa);
-                }
-            });
-            if (!filtered.isEmpty()) {
-                return ComputerUtilCard.getBestAI(filtered);
-            }
-            if (isOptional) {
-                return null;
-            }
-        }
-        return Iterables.getFirst(options, null);
     }
 }
